Finished Assignment
Add 50 Button adds 50 bananas, no matter how many are already in the
cart. Numbers them sequentially as well (if there are 50, it starts at
51 and goes to 100)

Clicking as URL goes to wikipidia banana page.

I made it so that "Select All/Select None" also worked on the "No Fruit
In Cart" cell (It was in an if statement, just copied it to the else
portion as well)